+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/transformer/JsonObjectToIrsPolicyTransformer.java
@@ -56,8 +56,8 @@ public class JsonObjectToIrsPolicyTransformer extends AbstractJsonLdTransformer<
             try {
                 final Object result = objectMapper.readerFor(Policy.class).readValue(v.asJsonObject().toString());
                 builder.permissions(((Policy) result).getPermissions());
-            } catch (JsonProcessingException e) {
-                throw new JsonParseException(e);
+            } catch (ClassCastException | JsonProcessingException e) {
+                throw new JsonParseException("Invalid policy", e);
             }
         });