+++ irs-policy-store/src/test/java/org/eclipse/tractusx/irs/policystore/testutil/PolicyStoreTestUtil.java
@@ -1,9 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022,2024
- *       2022: ZF Friedrichshafen AG
- *       2022: ISTOS GmbH
- *       2022,2024: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *       2022,2023: BOSCH AG
+ * Copyright (c) 2022,2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -21,18 +17,21 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.data;
+package org.eclipse.tractusx.irs.policystore.testutil;
 
-/**
- * Exception when parsing JSON
- */
-public class JsonParseException extends RuntimeException {
+import java.io.StringReader;
 
-    public JsonParseException(final String message, final Throwable throwable) {
-        super(message, throwable);
-    }
+import jakarta.json.Json;
+import jakarta.json.JsonObject;
+import jakarta.json.JsonReader;
 
-    public JsonParseException(final Throwable cause) {
-        super(cause);
+public class PolicyStoreTestUtil {
+
+    public static JsonObject toJsonObject(final String jsonString) {
+        final JsonObject jsonObject;
+        try (JsonReader jsonReader = Json.createReader(new StringReader(jsonString))) {
+            jsonObject = jsonReader.readObject();
+        }
+        return jsonObject;
     }
 }